Make handover fields optional in DeclarePreSubscriber
Made the handover-related fields optional, as the client does not have
this information. These fields will be populated by the client-attached
router.

An 'H' flag has been added to indicate whether the handover-related
fields are present. The DeclarePreSubscriber packet was also
refactored to group all handover fields into a contiguous block at the
end of the message, which is present only when the H flag is set.